Enable MockMvc with Mockito in JUnit 5 using a hacky workaround.
As there is not yet an official artifact for Mockito 2.x released that
adds a MockitoExtension (see [^1] and [^2]), use a custom one from [^3] for now.

This is urgently needed to enable unit testing of controllers WITHOUT the need
to write a @SpringBootTest which would involve starting a WebContext for the tests.
(That would be more in the direction of a integration test than necessary... See [^4])

Needs to be replaced with the official extension once released.

[^1]: https://github.com/mockito/mockito/issues/445
[^2]: https://github.com/mockito/mockito/issues/1221
[^3]: https://github.com/JeffreyFalgout/junit5-extensions/tree/master/mockito-extension
[^4]: https://thepracticaldeveloper.com/2017/07/31/guide-spring-boot-controller-tests