Add (not yet complete) controller AttributeAdminCtrl and unit test.
This relates to #56, which describes the metadata attribute requirement.
It solves this issue partially, as from the CRUD methods only CREATE and READ implemented and tested with a unit test.

Still missing are UPDATE and DELETE methods plus integration tests.

Updates the UML class diagram for package doc.attribute, too.