Refactor MongobeeConfig to be @SpringBootTest compatible.
Running future integration tests fail as the Mongobee configuration
was to static. It was not compatible with the usage of Embedded MongoDB
and random port assigning, as the port were mapped hard from the config.

By using @Autowire on MongoClient and MongoTemplate, the code is again
loosely coupled and starts working when using an inside-box integration
test with @SpringBootTest and @EnableAutoConfiguration to start the
embedded MongoDB.